backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
asyncpg==0.29.0
greenlet==3.0.3
aiosqlite==0.20.0
pytest==8.2.0
pytest-asyncio==0.23.7
openai==1.30.1
psycopg2-binary==2.9.9
